FlavorPage: user can add a review by clicking fab and reviews will update

+++ lib/FlavorPage/review.dart
@@ -1,4 +1,4 @@
-/// The [Review] class describes an ice cream reviews.
+/// The [Review] class describes an ice cream review.
 /// 
 /// Qualitative data
 /// 
@@ -18,7 +18,7 @@ class Review {
   String date;
   String text;
 
-  int reviewStars;
+  double reviewStars;
   int helpfulYes;
   int helpfulNo;
